backend_gateway/requirements.txt
circuitbreaker
django
djangorestframework
requests
